requirements.txt
Flask==3.0.0
python-dotenv==0.20.0
neo4j==5.12.0
fastapi==0.104.1
uvicorn==0.24.0

#python -m venv venv
#source venv/bin/activate 
#pip install -r requirements.txt
#uvicorn main:app --port 8081